split files by encoder / decoder
