+++ flif-dec.h
@@ -1,0 +1,6 @@
+#ifndef __FLIF_DEC_H__
+#define __FLIF_DEC_H__
+
+bool decode(const char* filename, Images &images, int quality, int scale);
+
+#endif
